% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIS.source.R
\name{RIS.source}
\alias{RIS.source}
\title{Source part of a R script}
\usage{
RIS.source(file, start, end, ...)
}
\arguments{
\item{file}{specify R script}

\item{start}{specifies start line}

\item{end}{specifies end line}
}
\description{
This function allows you to source part of a R script.
}
\examples{
RIS.source("test.R", start = 69, end = 420)
}
\keyword{RIS}
