% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIS.test.R
\name{RIS.test}
\alias{RIS.test}
\title{This is a function to test functionality}
\usage{
RIS.test()
}
\description{
Write RIS.test and see whether you get an output.
If yes, the package ist installed correctly.
}
\examples{
RIS.test()
}
\keyword{test}
