% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RIS.update.R
\name{RIS.update}
\alias{RIS.update}
\title{Update the RISpackage}
\usage{
RIS.update()
}
\description{
This function allows you to update the RISpackage from github.
}
\examples{
RIS.update()
}
\keyword{RISpackage}
\keyword{github}
\keyword{update}
